+++ app/admin/filter.rb
@@ -1,10 +1,13 @@
 ActiveAdmin.register Filter do
   actions :index, :show
 
+  remove_filter :card_order
+
   index do
     column :id
     column :created_at
     column :updated_at
+    actions
   end
 
   show do