+++ app/controllers/home_controller.rb
@@ -1,5 +1,0 @@
-class HomeController < ShopifyApp::AuthenticatedController
-  def index
-    render text: "Oops! If you get stuck here we messed up somehow :( Please let us know! support@touchcard.co"
-  end
-end