+++ spec/rails_helper.rb
@@ -43,7 +43,7 @@ RSpec.configure do |config|
   config.include StripeHelper, type: :controller
   # XXX these helpers should probably be intergrated better and selectively
   # loaded or loaded everywhere
-  config.include ShopifyHelper, type: :model
+  config.include ShopifyHelper
 
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false